Fixing windows build (#3)
* made covergage main-only.

* made covergage windows-only temp.

* trying minya/setup-cpp@v1.

* added CMAKE_TLS_VERIFY: "0" to build step to fix the cmake FetchContent for googletest

* added CMAKE_GENERATOR: "Unix Makefiles" to build step to build makefiles

* fixing dll path.

* fixing dll path #2.

* fixing dll path #3.

* fixing dll path #4.

* using default compiler on win.

* using default compiler on win #2

* using cl compiler on win

* configuring gtest in cmake to link statically the tests to the gtest lib.

* configuring gtest in cmake to link statically the tests to the gtest lib #2 + adding new gcc win build.

* fixing CI.

* fixing CI #2.

* fixing CI #3.

* modified Makefile to use pure cmake one-liners when possible.

* modified CI to remove use of make.

* fixed compiler options for MSVC.

* simplified cache key.

* implemented MSVC warnings.

* deactived error on warning (/WX) on windows builds because of too many errors.

* fixed  cache key.

* upgraded gtest to 1.17.0.

* fixed some MSVC reported errors.

* still fixing MSVC and cache.

* still fixing MSVC quirks.

* added tests.h to mutualize tests utils.

* added Magali to the acknowledgements.

* disabling msvc build since R is built using mingw (gcc)

* adding MAKE_GENERATOR: "Unix Makefiles" to disable ninja

* restoring the setup windows step

* adding lib build dir to path.

* fixed test compliation warnings.

* still fighting vs CI.

* still fighting vs CI #2.

* added save cache step. added GTEST_CATCH_EXCEPTIONS=0

* restricted the paths to be saved in the cache.

* restricted the paths to be saved in the cache#2.

* debugging fuzzy_rules_test.cpp on windows.

* fixed bug in fuzzy_rules_test.cpp and fixing cache in CI.

* added install of git via setup-cpp.

* still looking for git.

* still looking for git # 3.

* still looking for git # 4.

* still looking for git # 5.

* still looking for git # 6.

* still looking for git # 7. removed shell:bash from make build step

* still looking for git # 8. added shell:bash from report tests step

* still looking for git # 9. added new debug step

* still looking for git #10. had to force shell: cmd

* still looking for git #11. disable test reporter on windows

* still looking for git #12. disable test reporter on windows#2

* removing make install and cache.

* make install is required

* reactivated linux and macos

* reactivated linux and macos #2

* fixed junit.xml path

* fixing macos to macos-latest

* cleaninp up a bit

+++ src/bitarray.cpp
@@ -21,20 +21,20 @@ void BitArrayUtils::encode_number(int number, BitArray::iterator it, int nb_bits
 }
 
 void BitArrayUtils::randomize(BitArray& bits, RandomGenerator& rng) {
-    const int nb = bits.size();
+    const size_t nb = bits.size();
     vector<int> probs(nb);
     probs.resize(0);
-    rng.random(0, 1, nb, probs); // batch
+    rng.random(0, 1, int(nb), probs); // batch
 
-    for (int i = 0; i < nb; i++) {
-        bits[i] = probs[i];
+    for (size_t i = 0; i < nb; i++) {
+      bits[i] = probs[i];
     }
 }
 
 namespace fuzzy_coco {
   ostream& operator<<(ostream& out, const BitArray& bits) {
-      for (const auto& bit: bits)
-          out << noboolalpha << bit;
-      return out;
+    for (const auto& bit: bits)
+      out << noboolalpha << bit;
+    return out;
   }
 }
