+++ src/fuzzy_default_rule.cpp
@@ -35,7 +35,7 @@ vector<FuzzyDefaultRule> FuzzyDefaultRule::loadDefaultRules(const NamedList& des
 {
   vector<FuzzyDefaultRule> default_rules;
   default_rules.reserve(desc.size());
-  for (int i = 0; i < desc.size(); i++) 
+  for (size_t i = 0; i < desc.size(); i++) 
     default_rules.emplace_back(FuzzyDefaultRule::load(desc[i], db));
   return default_rules;
 }
